requirements.txt
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
dacite==1.8.1
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.3
fabric==3.2.2
h11==0.14.0
idna==3.4
invoke==2.2.0
marshmallow==3.20.1
marshmallow-dataclass==8.6.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.2
paramiko==3.3.1
psutil==5.9.6
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==2.0.6
webdriver-manager==4.0.1
wrapt==1.15.0
wsproto==1.2.0
python-dotenv==1.0.0
